+++ visualizations/points_actor.h
@@ -21,7 +21,7 @@ class PointsActor {
 
   vtkIdType AddPoint(double x, double y, double z);
 
-  template <int num_bits, int denom_bits>
+  template <size_t num_bits, size_t denom_bits>
   vtkIdType AddPoint(const HomoPoint3<num_bits, denom_bits>& p) {
     double w = p.w();
     return AddPoint(double(p.x()) / w, double(p.y()) / w, double(p.z()) / w);
@@ -29,7 +29,7 @@ class PointsActor {
 
   void SetPoint(vtkIdType index, double x, double y, double z);
 
-  template <int num_bits, int denom_bits>
+  template <size_t num_bits, size_t denom_bits>
   void SetPoint(vtkIdType index, const HomoPoint3<num_bits, denom_bits>& p) {
     double w = p.w();
     SetPoint(index, double(p.x()) / w, double(p.y()) / w, double(p.z()) / w);
